Refactor currency view with a VM
