+++ Travel KZH/Travel KZH/Views/Currency View/CurrencyView.swift
@@ -5,21 +5,18 @@
 //  Created by Joe Pham on 2022-08-29.
 //
 
-import MastercardOAuth1Signer
 import SwiftUI
 
 struct CurrencyView: View {
 	// MARK: Properties
-	@State private var rateInfo: CustomMastercardProperties?
-	@State private var rate = UserDefaults().getRate() ?? 0.0
-	@State private var rateDate = UserDefaults().getRateDate()
+	@StateObject
+	private var viewModel: CurrencyView.ViewModel
+	
+	init(viewModel: CurrencyView.ViewModel = .init()) {
+		_viewModel = StateObject(wrappedValue: viewModel)
+	}
 	
 	@FocusState private var customKZTAmountIsFocused: Bool
-	@State private var customKZTAmount: Double?
-	var convertedCHFAmount: Double { (customKZTAmount ?? 0) * rate }
-	
-	@State private var displayAlert = false
-	@State private var errorMessage: String?
 	
     var body: some View {
 		List {
@@ -30,9 +27,11 @@ struct CurrencyView: View {
 					HStack {
 						Text("₸")
 						
-						TextField("0.00",
-								  value: $customKZTAmount,
-								  format: .number) //format: .currency(code: "KZT"))
+						TextField(
+							"0.00",
+							value: $viewModel.customKZTAmount,
+							format: .number
+						) //format: .currency(code: "KZT"))
 						.keyboardType(.decimalPad)
 						.focused($customKZTAmountIsFocused)
 						
@@ -49,13 +48,20 @@ struct CurrencyView: View {
 					.cornerRadius(7)
 					
 					// KZT amount converted in CHF
-					Text(convertedCHFAmount, format: .currency(code: "CHF"))
-						.frame(minWidth: 0, maxWidth: .infinity, alignment: .trailing)
+					Text(
+						viewModel.convertedCHFAmount,
+						format: .currency(code: "CHF")
+					)
+					.frame(
+						minWidth: 0,
+						maxWidth: .infinity,
+						alignment: .trailing
+					)
 				}
 				.padding(.vertical, 4)
 				.font(.system(size: 17, weight: .regular, design: .rounded))
 			} header: {
-				if let info = getHeaderString() {
+				if let info = viewModel.getHeaderString() {
 					Text(info)
 				}
 			}
@@ -69,8 +75,8 @@ struct CurrencyView: View {
 						
 						Spacer()
 						
-						if rate != 0 {
-							Text(String(format: "CHF %.2f", (Double(amount) * rate)))
+						if viewModel.rate != 0 {
+							Text(String(format: "CHF %.2f", (Double(amount) * viewModel.rate)))
 						} else {
 							Text("CHF n/a")
 						}
@@ -78,13 +84,13 @@ struct CurrencyView: View {
 					.font(.system(size: 17, weight: .regular, design: .rounded))
 				}
 			} footer: {
-				Text(getFooterString())
+				Text(viewModel.getFooterString())
 			}
 		}
-		.alert("API Error", isPresented: $displayAlert) {
+		.alert("API Error", isPresented: $viewModel.displayAlert) {
 			Button("Continue") {}
 		} message: {
-			if let errorMessage = errorMessage {
+			if let errorMessage = viewModel.errorMessage {
 				Text(errorMessage)
 			}
 		}
@@ -92,13 +98,13 @@ struct CurrencyView: View {
 			customKZTAmountIsFocused.toggle()
 		}
 		.refreshable {
-			getCurrentRate()
+			viewModel.getCurrentRate()
 		}
 		.navigationTitle("Kazakhstan 🇰🇿")
 		//            .navigationBarTitle("Kazakhstan 🇰🇿", displayMode: .inline)
 		.toolbar {
 			Button {
-				getCurrentRate()
+				viewModel.getCurrentRate()
 			} label: {
 				Label("Refresh Rate", systemImage: "arrow.clockwise")
 			}
@@ -108,7 +114,7 @@ struct CurrencyView: View {
 				Button("Clear") {
 					// TODO: Remove focus works but not practical
 					customKZTAmountIsFocused.toggle()
-					customKZTAmount = nil
+					viewModel.customKZTAmount = nil
 				}
 				
 				Spacer()
@@ -122,106 +128,7 @@ struct CurrencyView: View {
 }
 
 private extension CurrencyView {
-	// MARK - Update rate provided by Mastercard API
-	func getCurrentRate() {
-		// TODO: - Import your certificate in this Xcode project
-		// TODO: - Change the name with the one from your certificate
-		let certificateName = "YOUR_CERTIFICATE_NAME_HERE"
-		let certificateExtension = "p12"
-		
-		if let certificatePath = Bundle.main.path(forResource: certificateName, ofType: certificateExtension) {
-			// Certificate path retrieved ✅
-			// TODO: - Change the password with the one from your certificate readme.txt
-			let keystorePassword = "YOUR_KEYSTORE_PASSWORD" // from certificate readme.txt
-			
-			if let signingKey = KeyProvider.loadPrivateKey(fromPath: certificatePath, keyPassword: keystorePassword) {
-				// Private Key loaded ✅
-				// TODO: - Edit the Consumer Key with the one from the Mastercard Developer Portal
-				let consumerKey = "YOUR_CONSUMER_KEY_HERE" // from developer portal
-				let url = getURL()
-				
-				do {
-					let header = try OAuth.authorizationHeader(forUri: url, method: "GET", payload: nil, consumerKey: consumerKey, signingPrivateKey: signingKey)
-					// OAuth Authorization Header created ✅
-					var urlRequest = URLRequest(url: url)
-					urlRequest.httpMethod = "GET"
-					urlRequest.allHTTPHeaderFields = [
-						"Authorization": header, // Add token in request header
-						"Accept": "application/json",
-						"Referer": "api.mastercard.com"
-					]
-					
-					
-					let task = URLSession.shared.dataTask(with: urlRequest) { data, response, error in
-						if let data = data {
-							if let response = try? JSONDecoder().decode(CustomMastercardProperties.self, from: data) {
-								DispatchQueue.main.async {
-									// Save locally/Update local rate
-									updateLocalData(rate: response.data.effectiveConversionRate, rateDate: response.requestDate)
-								}
-								
-								return
-							} else {
-								errorMessage = "Failed to decode JSON object."
-								displayAlert.toggle()
-							}
-						} else if let error = error {
-							errorMessage = "HTTP request failed: \(error)."
-							displayAlert.toggle()
-						}
-					}
-					
-					task.resume()
-				} catch {
-					errorMessage = "Failed to create the OAuth Authorization Header."
-					displayAlert.toggle()
-				}
-			} else {
-				errorMessage = "Failed to load the Private Key."
-				displayAlert.toggle()
-			}
-		} else {
-			errorMessage = "Failed to locate the project's certificate."
-			displayAlert.toggle()
-		}
-	}
 	
-	// MARK - Get and create Mastercard API URL
-	func getURL() -> URL {
-		let cardCurrency = "CHF"
-		let transactionCurrency = "KZT"
-		let date = Date.now.getFormattedDate()
-		
-		var urlString = "https://sandbox.api.mastercard.com/enhanced/settlement/currencyrate/subscribed/summary-rates"
-		urlString += "?rate_date=\(date)"
-		urlString += "&trans_curr=\(transactionCurrency)"
-		urlString += "&trans_amt=1"
-		urlString += "&crdhld_bill_curr=\(cardCurrency)"
-		
-		return URL(string: urlString)!
-	}
-	
-	// MARK - Get Footer Date
-	func getFooterString() -> String {
-		if let date = UserDefaults().getRateDate()?.toDate() {//rateDate?.toDate() {
-			let formattedDate = date.getReadableDate()
-			return "Last updated: \(formattedDate)"
-		}
-		
-		return "Pull to refresh the exchange rate."
-	}
-	
-	// MARK - Get Header Date
-	func getHeaderString() -> String? {
-		return rate != 0 ? "Exchange rate \(rate)" : nil
-	}
-	
-	func updateLocalData(rate: Double, rateDate: String) {
-		UserDefaults().saveRate(rate)
-		UserDefaults().saveRateDate(rateDate)
-		self.rate = rate
-		self.rateDate = rateDate
-	}
 }
 
 struct CurrencyView_Previews: PreviewProvider {
